pyproject.toml
[project]
name = "obsidian-api"
version = "0.1.0"
description = "FastAPI service for managing and searching an Obsidian vault."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=2.3.4",
    "python-frontmatter>=1.1.0",
    "uvicorn>=0.34.2",
    "python-dotenv>=1.1.0",
    "auth0-fastapi-api>=1.0.0b3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
]
